+++ src/Watashi.Shared/Constants/FileEntryTypes.cs
@@ -21,4 +21,5 @@ public static class SettingKeys
     public const string PasswordWarningDays = "PasswordWarningDays";
     public const string AgentMaxConcurrency = "AgentMaxConcurrency";
     public const string SessionIdleMinutes = "SessionIdleMinutes";
+    public const string AuditLogRetentionDays = "AuditLogRetentionDays";
 }